requirements.txt
appdirs==1.4.4
Arpeggio==2.0.0
asgiref==3.7.2
async-generator==1.10
attr==0.3.1
attrs==21.2.0
bitarray==2.8.2
bitmath==1.3.3.1
bitstring==4.1.2
bleach==6.0.0
blessed==1.17.12
bpylist2==4.1.1
certifi==2023.7.22
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
coremltools==5.2.0
cryptography==41.0.4
Django==4.2.7
docutils==0.19
h11==0.13.0
idna==2.10
importlib-metadata==6.1.0
jaraco.classes==3.2.3
jedi==0.18.2
keyring==23.13.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==8.14.0
mpmath==1.3.0
numpy==1.22.3
objexplore==1.6.3
opencv-python==4.5.5.64
osxmetadata==1.3.1
osxphotos==0.64.0
outcome==1.1.0
packaging==23.2
parso==0.8.3
pathvalidate==2.5.2
photoscript==0.3.0
Pillow==10.0.1
pkginfo==1.9.6
prompt-toolkit==3.0.38
protobuf==3.18.3
ptpython==3.0.23
py-applescript==1.0.3
pycparser==2.21
Pygments==2.15.0
pyobjc-core==9.0.1
pyobjc-framework-AppleScriptKit==9.0.1
pyobjc-framework-AppleScriptObjC==9.0.1
pyobjc-framework-AVFoundation==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-CoreAudio==9.0.1
pyobjc-framework-CoreMedia==9.0.1
pyobjc-framework-CoreML==9.0.1
pyobjc-framework-CoreServices==9.0.1
pyobjc-framework-FSEvents==9.0.1
pyobjc-framework-Metal==9.0.1
pyobjc-framework-Photos==9.0.1
pyobjc-framework-Quartz==9.0.1
pyobjc-framework-ScriptingBridge==9.0.1
pyobjc-framework-Vision==9.0.1
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
pytimeparse2==1.4.0
PyYAML==6.0
readme-renderer==37.3
requests==2.31.0
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.6.0
rich-theme-manager==0.11.0
scipy==1.10.0
selenium==3.141.0
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.4
strpdatetime==0.2.0
sympy==1.8
tenacity==8.2.2
textX==3.1.1
toml==0.10.2
tqdm==4.61.2
trio==0.20.0
trio-websocket==0.9.2
twine==4.0.2
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
wsproto==1.1.0
wurlitzer==3.0.3
xattr==0.10.1
xdg==5.1.1
zipp==3.15.0
